Delete dead code in eventHandler.js; unify serializers (fixes #24)
eventHandler.js exported six functions but only isEventEnabled had
callers; createClickHandler even set a clickEvent prop that does not
exist on the component. Meanwhile DeckGL.react.js carried its own newer
normalizePickInfo/serializeObject/safeClone.

Keep exactly one implementation: the live serializer moves into
eventHandler.js (exported), DeckGL.react.js imports it, and the five
dead exports (~170 lines) are gone. Unit tests for the util land with
the vitest setup in #25.

Verified via browser test: clicking a pickable ScatterplotLayer feature
delivers clickInfo (picked/layerId/object/coordinate) to a Dash
callback.
